# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 社内OSSカタログ / 納品用 OSS 管理 API
  version: 1.0.0-draft2
  description: |
    社内で利用・納品対象となる OSS を一元管理し、プロジェクト毎の利用状況と納品用一覧（SPDX / CSV 等エクスポート）を生成するための API 初稿 (Phase 1)。

    **本稿の位置づけ**
    - Phase 1 対象: OSSコンポーネント/バージョン CRUD, タグ, プロジェクト, プロジェクト利用 (Usage), スコープ判定, ポリシー参照, 監査ログ最小, エクスポート(プレースホルダ)
    - 未実装(将来): 脆弱性, SBOM Import, NOTICE 生成, ライセンス全文管理, **JWT リフレッシュ**, 監査詳細検索高度化

    **表記**
    - ID は原則 UUID v4 文字列
    - 日本語説明は実装後英語化検討
servers:
  - url: https://api.example.local/api/v1
    description: 本番 (社内)
  - url: https://dev.api.example.local/api/v1
    description: 開発環境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 互換用途の Basic 認証（将来廃止予定）
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: アクセストークン用 JWT（リフレッシュなし）
  parameters:
    PageParam:
      name: page
      in: query
      description: 1 始まりのページ番号
      schema: { type: integer, minimum: 1, default: 1 }
    SizeParam:
      name: size
      in: query
      description: 1ページ件数 (最大 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unauthorized:
      description: 認証失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 / RFC 7807 型エラー応答ボディ
      properties:
        type:
          type: string
          format: uri
          nullable: true
          description: 問題タイプ識別 URI（拡張分類用）
        title:
          type: string
          description: エラーの概要メッセージ（人間可読）
        status:
          type: integer
          description: HTTP ステータスコード
        detail:
          type: string
          nullable: true
          description: 追加詳細メッセージ
        instance:
          type: string
          format: uri
          nullable: true
          description: エラーが発生した具体的インスタンス URI（トレースID等）
        code:
          type: string
          nullable: true
          description: アプリケーション独自エラーコード
        errors:
          type: array
          description: フィールド単位バリデーションエラー配列
          items:
            type: object
            properties:
              field:
                type: string
                description: エラーが発生したフィールド名（JSON Pointer など）
              message:
                type: string
                description: フィールドに対するエラーメッセージ
      required: [title, status]

    # ---- JWT 関連追加 ----
...
    Tag:
      type: object
      description: OSSコンポーネントに付与する分類タグ
      properties:
        id: { type: string, format: uuid, description: タグ ID }
        name: { type: string, description: タグ名称（ユニーク） }
        createdAt: { type: string, format: date-time, description: 作成日時 }
      required: [id, name]
...
    TagCreateRequest:
      type: object
      description: タグ作成リクエスト
      required: [name]
      properties:
        name: { type: string, description: "タグ名称" }
...
    Role:
      type: string
      enum: [ADMIN, EDITOR, VIEWER]
      description: アクセス制御ロール（ADMIN=全権, EDITOR=編集可, VIEWER=参照のみ）
      x-enumDescriptions:
        - 管理者。ユーザー管理/ポリシー変更/全CRUDが可能
        - 編集者。OSS/プロジェクト/利用情報の作成・更新・削除が可能（ユーザー管理不可）
        - 閲覧者。読み取りのみ可能
    User:
      type: object
      description: システム利用ユーザー情報
      properties:
        id: { type: string, format: uuid, description: "ユーザーID" }
        username: { type: string, description: "ログインID（ユニーク）" }
        displayName: { type: string, description: "表示名" }
        email:
          {
            type: string,
            format: email,
            nullable: true,
            description: "メールアドレス",
          }
        roles:
          type: array
          description: 付与ロール一覧
          items: { $ref: "#/components/schemas/Role" }
        active: { type: boolean, description: "有効/無効フラグ" }
        createdAt: { type: string, format: date-time, description: "作成日時" }
        updatedAt: { type: string, format: date-time, description: "更新日時" }
      required: [id, username, roles, active, createdAt, updatedAt]
    UserCreateRequest:
      type: object
      description: ユーザー作成リクエスト
      required: [username, roles]
      properties:
        username: { type: string, description: "ログインID" }
        displayName: { type: string, description: "表示名" }
        email:
          {
            type: string,
            format: email,
            nullable: true,
            description: "メールアドレス",
          }
        password:
          {
            type: string,
            description: "初期パスワード（JWTでも登録時のみ使用可）",
          }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
          description: 付与するロール一覧
        active: { type: boolean, default: true, description: "有効化フラグ" }
    UserUpdateRequest:
      type: object
      description: ユーザー更新リクエスト（部分）
      properties:
        displayName: { type: string, description: "表示名" }
        email:
          {
            type: string,
            format: email,
            nullable: true,
            description: "メールアドレス",
          }
        password: { type: string, description: "パスワード再設定用（任意）" }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
          description: 付与ロール一覧（管理者のみ変更可）
        active: { type: boolean, description: "有効/無効切替" }
    PagedResult_User:
      type: object
      description: ユーザー一覧ページング結果
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
paths:
...
  /tags:
    get:
      tags: [Tags]
      summary: タグ一覧
      operationId: listTags
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Tags]
      summary: タグ作成
      operationId: createTag
      x-rolesAllowed: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tags/{tagId}:
    delete:
      tags: [Tags]
      summary: タグ削除
      operationId: deleteTag
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: tagId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /users:
    get:
      tags: [Users]
      summary: ユーザー一覧
      operationId: listUsers
      x-rolesAllowed: [ADMIN]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: username
          in: query
          schema: { type: string }
          description: 部分一致検索
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/Role" }
          description: ロールで絞り込み
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedResult_User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: ユーザー作成
      operationId: createUser
      x-rolesAllowed: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users/{userId}:
    get:
      tags: [Users]
      summary: ユーザー詳細
      operationId: getUser
      x-rolesAllowed: [ADMIN]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Users]
      summary: ユーザー更新
      operationId: updateUser
      x-rolesAllowed: [ADMIN]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Users]
      summary: ユーザー削除 (論理/物理は実装次第)
      operationId: deleteUser
      x-rolesAllowed: [ADMIN]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...